Merged PR 2021: Flutter/Android Logging
The flutter log level set in the configuration is passed to the Android component of the Flutter plugin and applied to the Android NxFit SDK. Currently the iOS plugin component ignores the log level set from Flutter.

The httpLogLevel set in the Flutter configuration is now passed into both Android and iOS native code via the configure message.

Also, Updated NxFit-SDK-for-iOS to 1.1.0, flutter  nxfit_sdk to 0.2.0 and Android NxFit SDK to 9.1.0.